fix: catch non existing sdp and svc ids on TiMOS
It seems like the SDP-ID and the SVC-ID are sometimes not set in the
SNMP response. Since the logic already checks if all 4 required values
are present, this commit just adds another check to check if the IDs are
present for the given OIDs. Additionally, it skips SAPs if the SVC-ID is
null.